Fixed Sql object not initializing default values properly. Fixed Sql_Free producing timer deallocation errors when called without first establishing a database connection.
git-svn-id: https://svn.code.sf.net/p/rathena/svn/trunk@13011 54d463be-8e91-2dee-dedb-b68131a5f0ec

+++ src/common/sql.c
@@ -67,9 +67,14 @@ struct SqlStmt
 Sql* Sql_Malloc(void)
 {
 	Sql* self;
+
 	CREATE(self, Sql, 1);
 	mysql_init(&self->handle);
 	StringBuf_Init(&self->buf);
+	self->lengths = NULL;
+	self->result = NULL;
+	self->keepalive = INVALID_TIMER;
+
 	return self;
 }
 
@@ -406,7 +411,7 @@ void Sql_Free(Sql* self)
 	{
 		Sql_FreeResult(self);
 		StringBuf_Destroy(&self->buf);
-		delete_timer(self->keepalive, Sql_P_KeepaliveTimer);
+		if( self->keepalive != INVALID_TIMER ) delete_timer(self->keepalive, Sql_P_KeepaliveTimer);
 		aFree(self);
 	}
 }
